actuarial-calc-engine-main/actuarial-calc-engine-main/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "actuarial-calculation-engine"
version = "0.1.0"
description = "A modular Python engine for actuarial calculations with Excel/PostgreSQL input and multiple output formats"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "pytest>=7.4.0"
]
authors = [
    {name = "Special project team"}
]

[project.scripts]
actuarial-engine = "src.actuarial_engine:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"